requirements.txt
pypresence>=4.3.0
pystray>=0.19.0
Pillow>=10.0.0
